Add Jenkins CI and automated tests for Debian
Whenever a PR is created in the dev or master branch, a Jenkins job will
build each Dockerfile in the repo and run tests on the Debian images.
The tests perform a simple smoke test on the images.

The plan is to later introduce testing for the Windows Server images
also.